Add files via upload
First draft of program. All functionality works, however the code is still clunky, and the UI is VERY simplistic. 

I'm hoping to add a "image preview" and the ability to choose what size/percentage you want the watermarks to be. Similar to how Lightroom's watermark preview works. However it does all the basic things it needs to do currently.